Move Result to model instead of util
