+++ data/src/main/java/com/github/leuludyha/data/repository/datasource/LibraryRemoteDataSource.kt
@@ -1,8 +1,7 @@
 package com.github.leuludyha.data.repository.datasource
 
 import com.github.leuludyha.domain.model.Search
-import com.github.leuludyha.domain.util.Result
-import retrofit2.Response
+import com.github.leuludyha.domain.model.Result
 
 interface LibraryRemoteDataSource {
     suspend fun search(query: String): Result<Search>
