+++ domain/src/main/java/com/github/leuludyha/domain/repository/LibraryRepository.kt
@@ -1,7 +1,7 @@
 package com.github.leuludyha.domain.repository
 
 import com.github.leuludyha.domain.model.Search
-import com.github.leuludyha.domain.util.Result
+import com.github.leuludyha.domain.model.Result
 
 interface LibraryRepository {
     suspend fun search(query: String): Result<Search>
